// src/interfaces/CadastroDeProduto.ts
export default interface CadastroDeProduto {
    nome?: string;
    categoria?: string;
    quantidade?: string;
    tamanho?: string;
    precoUnitario?: string;
    precoUnitarioVenda?: string;
    data?: string;
    teste?: string;
    marca?: string;
    modelo?: string;
    garantia?: string;
    pesoOuVolume?: string;
    validade?: string;
    cor?: string;
};